+++ rs-react-app/src/components/pages/SearchResults/ErrorBoundary.tsx
