+++ rs-react-app/src/components/pages/SearchResults/ResultsFieldFn.tsx
