+++ rs-react-app/src/components/pages/SearchResults/SearchBarFn.tsx
